Consolidate execution settings and validation into the orchestrator
