+++ src/rust/cli/src/execution/mod.rs
@@ -1,7 +1,5 @@
 mod authorization;
 mod orchestrator;
-mod settings;
-mod validation;
 
 pub use orchestrator::{execute, execute_with_observer, should_attempt_azure_auth};
